+++ src/index.ts
@@ -46,5 +46,8 @@ export * from './networking/WMTNetworking'
 export * from './WMTPlatformUtils'
 export * from './WMTSDKVersion'
 
+// UTILS
+export * from './utils/WMTAnyObject'
+
 // PWA MTOKEN INSTANTIATION
 export * from './PWAExtension'